fix purge checkpoint to keep the latest snapshot index

+++ rockredis/rockredis.go
@@ -120,7 +120,7 @@ func GetLatestCheckpoint(checkpointDir string, skipN int, matchFunc func(string)
 	return ""
 }
 
-func purgeOldCheckpoint(keepNum int, checkpointDir string) {
+func purgeOldCheckpoint(keepNum int, checkpointDir string, latestSnapIndex uint64) {
 	defer func() {
 		if e := recover(); e != nil {
 			dbLog.Infof("purge old checkpoint failed: %v", e)
@@ -134,6 +134,19 @@ func purgeOldCheckpoint(keepNum int, checkpointDir string) {
 		sortedNameList := CheckpointSortNames(checkpointList)
 		sort.Sort(sortedNameList)
 		for i := 0; i < len(sortedNameList)-keepNum; i++ {
+			fn := path.Base(sortedNameList[i+keepNum])
+			subs := strings.Split(fn, "-")
+			if len(subs) != 2 {
+				continue
+			}
+			sindex, err := strconv.ParseUint(subs[1], 16, 64)
+			if err != nil {
+				dbLog.Infof("checkpoint name index invalid: %v, %v", subs, err.Error())
+				continue
+			}
+			if sindex >= latestSnapIndex {
+				break
+			}
 			os.RemoveAll(sortedNameList[i])
 			dbLog.Infof("clean checkpoint : %v", sortedNameList[i])
 		}
@@ -158,6 +171,7 @@ type RockDB struct {
 	hllCache          *hllCache
 	stopping          int32
 	engOpened         int32
+	latestSnapIndex   uint64
 }
 
 func OpenRockDB(cfg *RockRedisDBConfig) (*RockDB, error) {
@@ -221,6 +235,10 @@ func (r *RockDB) CheckExpiredData(buffer common.ExpiredDataBuffer, stop chan str
 		return fmt.Errorf("can not check expired data at the expiration-policy:%d", r.cfg.ExpirationPolicy)
 	}
 	return r.expiration.check(buffer, stop)
+}
+
+func (r *RockDB) SetLatestSnapIndex(i uint64) {
+	atomic.StoreUint64(&r.latestSnapIndex, i)
 }
 
 func (r *RockDB) GetBackupBase() string {
@@ -664,8 +682,9 @@ func (r *RockDB) backupLoop() {
 				if r.cfg.KeepBackup > 0 {
 					keepNum = r.cfg.KeepBackup
 				}
-				purgeOldCheckpoint(keepNum, r.GetBackupDir())
-				purgeOldCheckpoint(MaxRemoteCheckpointNum, r.GetBackupDirForRemote())
+				// avoid purge the checkpoint in the raft snapshot
+				purgeOldCheckpoint(keepNum, r.GetBackupDir(), atomic.LoadUint64(&r.latestSnapIndex))
+				purgeOldCheckpoint(MaxRemoteCheckpointNum, r.GetBackupDirForRemote(), math.MaxUint64-1)
 				r.checkpointDirLock.Unlock()
 			}()
 		case <-r.quit:
@@ -909,8 +928,8 @@ func (r *RockDB) restoreFromPath(backupDir string, term uint64, index uint64) er
 		if r.cfg.KeepBackup > 0 {
 			keepNum = r.cfg.KeepBackup
 		}
-		purgeOldCheckpoint(keepNum, r.GetBackupDir())
-		purgeOldCheckpoint(MaxRemoteCheckpointNum, r.GetBackupDirForRemote())
+		purgeOldCheckpoint(keepNum, r.GetBackupDir(), atomic.LoadUint64(&r.latestSnapIndex))
+		purgeOldCheckpoint(MaxRemoteCheckpointNum, r.GetBackupDirForRemote(), math.MaxUint64-1)
 	}
 	return err
 }
